feat: add support for extracting private node agent parameters from Helm commands
