+++ configurations/system/tests_cases/runtime_tests.py
@@ -59,7 +59,8 @@ class RuntimeTests(object):
             name=inspect.currentframe().f_code.co_name,
             test_obj=IncidentResponse,
             deployments=join(DEFAULT_DEPLOYMENT_PATH, "redis_sleep_long"),
-            tests=["ApplyNetworkPolicy", "ApplySeccompProfile", "KillProcess"],
+            tests=["ApplyNetworkPolicy", "ApplySeccompProfile"],
+            with_private_node_agent=False,
             # create_test_tenant=True,
         )
